backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.6.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# LLM & RAG
openai==1.12.0
llama-index==0.10.12
langchain==0.1.6
langchain-openai==0.0.5

# Vector Database
chromadb==0.4.18

# Document Processing
beautifulsoup4==4.12.3
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.4
markdown==3.5.2

# Confluence (for later when accessing private Confluence)
atlassian-python-api==3.41.0

# GitHub & Git
PyGithub==2.1.1
GitPython==3.1.41

# Utilities
loguru==0.7.2
tenacity==8.2.3
httpx==0.25.0
packaging==23.2  # Fix langchain-core conflict

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Caching (optional - for conversation memory)
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
